requirements.txt
fastapi[standard]~=0.115.12
pydantic~=2.10.6
pydantic-settings
paramiko
scp
pymavlink
numpy
opencv-python~=4.11.0
Pillow